scriptkill-store-nosql/requirements.txt
flask==2.3.0
flask-cors==4.0.0
PyJWT==2.8.0
pymongo==4.15.0
redis==6.4.0
